pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "grpo-strict-generation"
version = "0.1.0"
description = "Align a Small LLM with GRPO for Strict Code/JSON Generation (Track 23)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"

dependencies = [
    # Core ML
    "torch",
    "transformers",
    "trl",
    "peft",
    "datasets",
    "accelerate",
    "bitsandbytes",
    # Data & Eval
    "numpy",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    # Infra
    "pyyaml",
    "tensorboard",
    "wandb",
    "python-dotenv",
    "tqdm",
    "jupyter",
    "ipywidgets",
]

[project.optional-dependencies]
dev = [
    "isort",
    "black[jupyter]",
    "ruff",
    "pytest",
]
fast = [
    "unsloth",
    "vllm",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-include = ["*.ipynb"]
exclude = ["notebooks/reference/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E402"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["F821", "F401"]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
